requirements.txt
fastapi
uvicorn
huggingface_hub
python-multipart
python-dotenv
openai
websockets
httpx
